clean legacy code
