+++ pyproject.toml
@@ -7,7 +7,6 @@ authors = [{name = "Jorge Rivera", email = "jorge.rivera@one.org"}]
 requires-python = ">=3.10, <4.0"
 dependencies = [
     "pandas >=2",
-    "bblocks >=1",
     "requests >=2",
     "lxml >=5",
     "pyarrow >=14, <19",
